refactor and added comments

+++ src/main/java/com/example/movieapi/service/UserService.java
@@ -10,7 +10,9 @@ import java.util.Set;
 import com.example.movieapi.models.ApplicationUser;
 import com.example.movieapi.models.Movie;
 import com.example.movieapi.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -20,16 +22,21 @@ import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+/**
+ * This class is the service for the user CRUD
+ * It is responsible for handling the business logic for the user routes
+ */
+
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
+@Transactional
 public class UserService implements UserDetailsService {
 
-    @Autowired
-    private PasswordEncoder encoder;
+    private final PasswordEncoder encoder;
 
-    @Autowired
-    private UserRepository userRepository;
+    private final UserRepository userRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
@@ -50,11 +57,11 @@ public class UserService implements UserDetailsService {
                 return ResponseEntity.ok(userRepository.findById(id));
             } else {
                 log.info("User not found");
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.status(204).build();
             }
         } catch (Exception e){
             log.info("Error getting user");
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(400).build();
         }
     }
 
@@ -63,30 +70,34 @@ public class UserService implements UserDetailsService {
             List<ApplicationUser> users = userRepository.findAll();
             if(users.isEmpty()) {
                 log.info("No users found");
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.status(204).build();
             } else {
                 log.info("Users found");
                 return ResponseEntity.ok(users);
             }
         } catch (Exception e) {
             log.info("Error getting users");
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(400).build();
         }
     }
 
     public ResponseEntity<String> deleteUser(Long id) {
         try {
+            if(id == 1){
+                log.info("Cannot delete admin");
+                return ResponseEntity.status(400).body("Cannot delete admin");
+            }
             if (userRepository.existsById(id)) {
                 userRepository.deleteById(id);
                 log.info("User deleted successfully");
                 return ResponseEntity.ok("User deleted successfully");
             } else {
                 log.info("User not found");
-                return ResponseEntity.notFound().build();
+                return ResponseEntity.status(204).build();
             }
         } catch (Exception e) {
             log.info("Error deleting User");
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(400).build();
         }
     }
 
@@ -100,7 +111,7 @@ public class UserService implements UserDetailsService {
             return ResponseEntity.ok(userRepository.save(userToUpdate));
         } catch (Exception e) {
             log.info("Error updating user");
-            return ResponseEntity.badRequest().build();
+            return ResponseEntity.status(400).build();
         }
     }
 }
